requirements.txt
pyglm
pyglet
pysdl2
pysdl2-dll
glfw
rsxpy
rvr
imgui
stbi-py